requirements.txt
aniso8601==9.0.1
asgiref==3.4.1
attrs==21.2.0
Automat==20.2.0
backports.entry-points-selectable==1.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.8
click==8.0.3
colorama==0.4.4
constantly==15.1.0
crochet==2.0.0
cryptography==35.0.0
cssselect==1.1.0
distlib==0.3.3
Django==3.2.9
django-cors-headers==3.10.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
filelock==3.1.0
Flask==2.0.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
h2==3.2.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
itemadapter==0.4.0
itemloaders==1.0.4
itsdangerous==2.0.1
Jinja2==3.0.3
jmespath==0.10.0
joblib==1.1.0
lxml==4.6.4
MarkupSafe==2.0.1
numpy==1.21.3
pandas==1.3.4
parsel==1.6.0
platformdirs==2.4.0
priority==1.3.0
Protego==0.1.16
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.5
PyJWT==1.7.1
pyOpenSSL==21.0.0
python-dateutil==2.8.2
pytz==2021.3
queuelib==1.6.2
requests==2.26.0
scikit-learn==1.0.1
scipy==1.7.3
Scrapy==2.5.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
threadpoolctl==3.0.0
Twisted==21.7.0
typing_extensions==4.0.0
urllib3==1.26.7
virtualenv==20.8.1
w3lib==1.22.0
Werkzeug==2.0.2
wrapt==1.13.3
zope.interface==5.4.0
